requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
asyncpg==0.29.0
aiosqlite==0.19.0
python-dotenv==1.0.1
APScheduler==3.10.4
python-dateutil==2.9.0.post0
aiohttp==3.9.1

